tg_back/requirements.txt
MySQLdb
PyJWT
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.4
simplejson==3.16.0
six==1.12.0
Werkzeug==0.16.0
